Merged test code into overall app
Good progress has been made. Writing data to arduino works well. We can
read data from the arduino but it is very unresponsive so we need to
look into a better way to do that. Work on integrating DAC and ADC into
overall app design. Code only works when run in a loop but it would be
very nice if it didn't have to be in a loop.