Reading and Writing with arduino works!
Terminator for strings in the serial port is LF(so the newline character must end all strings that are being exchanged). Arduino must manually print a newline character after every string it sends. Arduino also needs to print an initial message, otherwise the program will have a long delay at the start. To avoid having the entire MATLAB code in a loop, we need to create the arduino serial object and then delete it once the data has been exchanged.